Fix book reading progress and settings persistence
- Changed to save progress/settings only on app background or VC dismissal
- Fixed database schema by adding missing columns via migrations
- Fixed ttsSpeed type mismatch (Float to Double)
- Fixed chapter tracking when scrolling between chapters
- Updated updateBookReadingPreferences to use save (upsert) instead of update only
- Removed debug logging

This significantly reduces database writes and ensures reading position is properly restored when reopening books.

+++ aotd/Models/Book.swift
@@ -307,7 +307,7 @@ struct BookReadingPreferences: Codable, FetchableRecord, PersistableRecord {
     var scrollPosition: Double
     var brightness: Double
     var autoScrollSpeed: Double?
-    var ttsSpeed: Float
+    var ttsSpeed: Double
     var ttsVoice: String?
     var textAlignment: String // left, center, right, justified
     var marginSize: Double // 0-50 points
